git: Split hfs and ntfs into separate files
